+++ src/config/DBConnection.java
@@ -15,7 +15,6 @@ public class DBConnection {
             Class.forName("com.mysql.cj.jdbc.Driver");
         } catch (ClassNotFoundException e) {
             System.err.println("Connection Not Found!");
-
         }
     }